storybook: move time consts in common folder
