+++ front/ui/storybook/stories/ui-charts/common/const.ts
@@ -1,0 +1,4 @@
+// Here are some helpers to write code about time in ms that is humanly readable:
+export const SECOND = 1000;
+export const MINUTE = 60 * SECOND;
+export const HOUR = 60 * MINUTE;
