+++ front/ui/storybook/stories/ui-charts/spaceTimeChart/additional-data.stories.tsx
@@ -14,7 +14,7 @@ import cx from 'classnames';
 import { clamp, inRange } from 'lodash';
 
 import { MouseTracker } from './helpers/components';
-import { AMBIANT_A10, ERROR_30, ERROR_60, HOUR, KILOMETER, MINUTE } from './helpers/consts';
+import { AMBIANT_A10, ERROR_30, ERROR_60, KILOMETER } from './helpers/consts';
 import { OPERATIONAL_POINTS, PATHS, START_DATE } from './helpers/paths';
 import {
   MAX_X_ZOOM,
@@ -25,9 +25,9 @@ import {
   Y_ZOOM_LEVEL,
   getDiff,
 } from './helpers/utils';
-
 import '@osrd-project/ui-charts/dist/theme.css';
 import '@osrd-project/ui-core/dist/theme.css';
+import { MINUTE, HOUR } from '../common/const';
 
 const MONO_TRACK_SPACES = [
   { from: 6 * KILOMETER, to: 24 * KILOMETER },
