+++ front/ui/storybook/stories/ui-charts/spaceTimeChart/helpers/consts.ts
@@ -14,10 +14,5 @@ export const OCCUPANCY_FREE = '#CAEDDB';
 export const OCCUPANCY_SEMAPHORE = '#FFD4D8';
 export const OCCUPANCY_WARNING = '#FFEABF';
 
-// Here are some helpers to write code about time in ms that is humanly readable:
-export const SECOND = 1000;
-export const MINUTE = 60 * SECOND;
-export const HOUR = 60 * MINUTE;
-
 // Same for distances in millimeters:
 export const KILOMETER = 1000000;
