+++ front/ui/storybook/stories/ui-charts/spaceTimeChart/layers.stories.tsx
@@ -14,16 +14,15 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 
 import {
   KILOMETER,
-  MINUTE,
   OCCUPANCY_FREE,
   OCCUPANCY_SEMAPHORE,
   OCCUPANCY_WARNING,
 } from './helpers/consts';
 import { OPERATIONAL_POINTS, PATHS, START_DATE } from './helpers/paths';
 import { X_ZOOM_LEVEL, Y_ZOOM_LEVEL } from './helpers/utils';
-
 import '@osrd-project/ui-core/dist/theme.css';
 import '@osrd-project/ui-charts/dist/theme.css';
+import { MINUTE } from '../common/const';
 
 const CONFLICTS = [
   {
